fix: respect BUILD_DIR in Makefile

+++ Makefile
@@ -1,10 +1,9 @@
-BUILD_DIR ?= $(shell mktemp -d)
-
 .PHONY: install install-dev test update clean
 
 install:
-	@BUILD_DIR_TMP=$$(mktemp -d); \
-	trap 'rm -rf "$$BUILD_DIR_TMP"; cd quickjs-x && $(MAKE) clean-all' EXIT; \
+	@BUILD_DIR_TMP=$${BUILD_DIR:-$$(mktemp -d)}; \
+	[ -n "$(BUILD_DIR)" ] || trap 'rm -rf "$$BUILD_DIR_TMP"; cd quickjs-x && $(MAKE) clean-all' EXIT; \
+	mkdir -p $$BUILD_DIR_TMP; \
 	$(MAKE) -C quickjs-x BIN_DIR=$$BUILD_DIR_TMP/quickjs-x/bin $$BUILD_DIR_TMP/quickjs-x/bin/qjsx && \
 	BUILD_DIR=$$BUILD_DIR_TMP sh install.sh
